ashell: rename shell related files to use 'term' prefix to prepare introducing the 'ide' prefixed ones
