+++ src/ashell/Makefile
@@ -13,16 +13,16 @@ ZEPHYRINCLUDE += -I$(ZEPHYR_BASE)/samples/subsys/usb/webusb/src \
                  -I${ZEPHYR_BASE}/subsys/usb/class \
                  -I$(ZEPHYR_BASE)/include/usb
 obj-y += ../../deps/zephyr/samples/subsys/usb/webusb/src/webusb_serial.o
-obj-y += webusb-handler.o
+obj-y += term-webusb.o
 endif
 
 obj-y += file-utils.o
 obj-y += jerry-code.o
 
-obj-y += ihex-handler.o
+obj-y += term-ihex.o
 
-obj-y += comms-uart.o
-obj-y += shell-state.o
+obj-y += term-uart.o
+obj-y += term-cmd.o
 
 obj-y += ../../deps/ihex/kk_ihex_read.o
