--- src/ashell/comms-shell.c
@@ -1,2 +1,0 @@
-// Copyright (c) 2016-2017, Intel Corporation.
-
