--- src/ashell/comms-shell.h
@@ -1,6 +1,0 @@
-// Copyright (c) 2016-2017, Intel Corporation.
-
-#ifndef __comms_shell_h__
-#define __comms_shell_h__
-
-#endif  // __comms_shell_h__
